Add support for ign serve token rotation
This PR enables the token Secret controller to automatically delete token Secrets generated by the NodePool controller when the ttl is met by a secret.
The token Secret controller controller uses now a finalizer to ensure a token is removed from the payload cache when the token Secret is removed.
When that happens the NodePool controller reconciles and create a new token Secret.
Then it reconciles the userData Secret in place with the new generated token so new machines can join the cluster successfully. Therefore the userdata Secret is mutable.