Destroy cluster: Better deal with invalid credentials file
I accidentelly specified an invalid credentials file to the `destroy
cluster` command which resulted in being stuck in `Destroying
infrastructure` for four minutes and then telling me there was an error
loading the file and that it will retry. Looking at the code, this
happens because:
* We use `RetryUntil` which always waits Interval time before starting
* In the DNS zone cleanup we always retry all errors with an exponential
  backoff starting with one second, a 1.5 factor and 10 tries - For both
  public and private zone

This change fixes that by adding a function to check for credential load
errors and not retry them anywhere. It also replaces the wait.RetryUntil
with wait.RetryImmediateUntil to prevent the initial wait which we don't
need.