Update controller-runtime to 0.10.1
This updates controller-runtime (and with it the k8s.io/ dependencies)
to 0.10.1. The main motivation is that I want to start collecting
clientsside metrics, however there is currently `rest_client_request_latency_seconds`
which includes the full request URL and thus has a huge cardinality. It
got removed in the 0.10 release of controller-runtime.