create cluster: automatically create ssh key if one is not provided
Currently if the user does not supply a SSH key when creating a cluster,
no SSH key is used when igniting workers. If then there is a problem
bringing up those workers, there is no good way to ssh to those workers.
This change adds code to generate a SSH key when one is not provided.
The private key is stored in the same secret if later it's needed to SSH
into the worker machines.