Make reconciliation of user-manifest-konnectivity-agent-secret CM idempotent
That configmap holds a secret with tls credentials. Unfortunately, we
never unmarshal its content into the secret we later on use to check if
we have to update it, resulting in us always updating it.

This was made a bit better by the fact that we also didn't watch
configmaps, which is why it didn't start a loop.