Prevent no-op updates of Kube objects
We are currently doing a lot of Update calls for Kube objects that don't
actually change anything and happen because the server defaults fields
and our code resets them to be empty. We prevent this by implementing a
custom CreateOrUpdate that will copy fields that get defaulted from the
object on the server to the locally-modified object, if unset in the
latter.

Additinally, we will log if we detect multiple updates to the same
object without an attempted update that resulted in a no-op. We then
verify in the e2e tests that we don't log this to detect regressions.

Ref https://issues.redhat.com/browse/HOSTEDCP-237