AWS: Fix recognizing existing intances for a machine
The aws capa machine reconciler uses a couple of filters[0] to find
existing instances for a machine. Among these is one for the VPC. The
VPC ID in turn is populated from the AWSCluster object[1].

Unfortunatelly, we never set the VPC ID on the AWSCluster object,
resulting in us never recognizing existing instances and in turn always
creating a new instance if the awsmachine's .spec.providerID field is
empty or invalid.

[0]: https://github.com/kubernetes-sigs/cluster-api-provider-aws/blob/aa68f949dc496116155c73fd61ea5760f493f6cb/pkg/cloud/services/ec2/instances.go#L50-L55
[1]: https://github.com/kubernetes-sigs/cluster-api-provider-aws/blob/aa68f949dc496116155c73fd61ea5760f493f6cb/pkg/cloud/scope/cluster.go#L107