Fix local hypershift operator CPO flag regression
This commit fixes a regression introduced in f82b00f6a563d1439b0d495872dd35fa9e833af1
which makes it impossible to run the hypershift operator locally without having the
operator deployment installed in the cluster. The user should be able to override the
CPO fallback image with a flag like before, and deployment image lookups should only
be used when running in-cluster (i.e. without the flag).