imageregistry config: Don't permanently create new secret
Currently, we permanently update the configmap that contains the
imageregistry config, because it contains a httpSecret that is
re-generated all the time. Avoiding that in the current code structure
is extremely difficult so instead a new resources controller is added to
the hosted-controlplane operator. That controller currently reconciles
only the imageregistry config but should eventually reconcile everything
that is currently in the manifest bootstrapper.

The perma-reconciling happened to only show up as a lot of updates of
the configmap but didn't show up in the guest cluster as the configmap
is only applied once and didn't cause a loop as the configmap isn't
watched.