Convert mco script to template, conditionally include mdns publisher image
In 4.10, there is no longer an mdns publisher image and the argument has
been removed from the mco bootstrap command, resulting in an error. This
PR converts the script to run the mco bootstrap command into a go
template and excludes the argument if the mdns publisher image is not
present.