KAS and Konnektivity svc: Fix idempotency
I noticed that the KAS and Konnektivity service are always updated twice
when restarting the CPO. It turns out this happens because:
* They reference a package-level variable for their selector
* The same variable is used in the Deployments selector
* The deployments label selector is manipulated through the
  DeploymentConfigs by setting the colocation label, which ends up
  propagating back to the package-level variable
* Because the service is created before the deployment, we always remove
  that label and then re-add it.

This happened to not be noticed by the update loop detector because on
cluster create, we first create those services and then wait for the
infra to be ready by returning in the Reconcile. This means that on the
next iteration we go through the service upsert code again and do a
no-op update there, resulting in the loop detector being disabled for
these objects.