Openshift APIServer: Trust the service signer CA
The openshift apiserver needs to connect to the in-cluster registry
whose serving certificate is signed by the service signer CA, thus it
needs to trust the service signer CA. Because there is no config option
for trusted CAs, we need to override the global CA bundle. Since the
global CA bundle is under /etc we can not directly write to it. Instead,
we use an initcontainer that copies the CA bundle onto an emptyDir
volume, appends the service signer ca and then mount that empty dir into
the main openshift apiserver container.

Additionally:
* Rename the kube-controller-manager configmap that only contains the
  service signer CA to more clearly reflect this
* Make the cmca controller also annotate the podspec template of the
  openshift apiserver with the CA hash, so that it gets rotated when the
  CA changes