Verify if cluster capabilities match the config
Use `ValidConfiguration` condition for the hosted cluster controller. If
the capabilities check fails, this condition is set to false and
reconciliation will be blocked.

Introduce a new `ValidHostedControlPlaneConfiguration` condition which
will be used by the controlplane controller to indicate configuration
issues.

In case of invalid configurations the condition is bubbled up to the
hosted cluster controller.

Signed-off-by: Roman Mohr <rmohr@redhat.com>