AWS: Fix creation of pre-provisioned PVs
There seems to be an admission webhook in the kube-apiserver that checks
if the volume of pre-provisioned PVs exist and that currently fails with
authentication, as there are no AWS credentials. Mount the kcm creds on
AWS to fix that.

After this change, all `[sig-storage] In-tree Volumes [Driver: aws]`
tests in the conformance suite pass.

Ref https://issues.redhat.com/browse/HOSTEDCP-257