Improve API reference docs
This commit updates API documentation on the Go types and also improves
the generated API documentation to include only the public types, as well
as other minor small fixes for readability.

Note that the `+gencopy` comments added to the API types are to work around
an issue[1] in the upstream docs generation code which doesn't yet understand
the `+kubebuilder:object:root` tag. This is necessary to make the generator
understand which types are "public" (e.g. HostedCluster and NodePool).

[1] https://github.com/ahmetb/gen-crd-api-reference-docs/pull/26