Replace MCS config templates with go reconciliation code
Moves code that parses and reconciles global config to a separate
package under support. The reason is that we need the same code to be
used by the hosted-cluster-config-operator.

Removes static templates related to the generation of configuration for
the MCO/MCS. They are replaced with go reconciliation code under the
'mcs' package.

Removes additional mcs kubeconfig secret and simply points the ignition
provider to use the existing bootstrap secret as generated by the PKI
code.