AWS: Fix volumes by passing cloud provider creds to KAS
The KAS needs to cloud provider creds as it communicates with the cloud
provider for admitting them. Updating the KAS was forgotten when
refactoring the code to use STS. As cluster creation will fail if the
KAS has an initcontainer that requires the KAS to be running, the
token-minter was changed to be a sidecar instead of an initcontainer.

Also, the corresponding code was refactored to be shared between the KAS
and the KCM to avoid this kind of mistake in the future and de-duplicate
it.

Ref https://issues.redhat.com/browse/HOSTEDCP-257