Move the apiserver-proxy into a DaemonSet
Running it as static pod on the node results in the Kubelet adding a
keyless noExecute toleration, which makes the `[sig arch] ensure control
plane operators do not make themselves unevictable` conformance test
fail.

Make it a DaemonSet instead, which also has the advantage that we can
change it without re-provisioning the node.