Hypershift operator: Ensure created objects are annotated to trigger reconciliation
The hypershift operator creates a number of resources per hostedcluster
and changes to these trigger a reconciliation of the corresponding
hostedcluster. The mapping is done through an annotation.

Unfortunately, we do not set the annotation in many cases which means
that external changes do not get corrected by the Hypershift operator.

This change makes us automically insert the annotation. It went through
several rounds before I was happy with the result:
1. Wrap the client and just check and error: This works but requires to
   update a massive amount of code to insert the annotation everywhere
   and plug the required info through
2. Wrap the client to automatically insert the annotation: This mostly
   works, but if someone ever removes the annotaiton, we will not re-add
   it because CreateOrUpdate doesn't see the diff, as it happens on a
   layer below
3. Add a CreateOrUpdate wrapper that gets constructed with the
   reconcile.Request and wraps the mutateFN: This makes CreateOrUpdate
   automatically insert the annotation and see the diff created by this,
   if any. The wrapper is tiny so this seems pretty elegant to me.