Default MachineDeployment replicas to 0 if they are nil
Before this PR MachineDeployments with nil replicas i.e brand new MachineDeployments were temporary defaulted to 1 when a brand new NodePool replicas was nil. The immediate next call to setMachineDeploymentReplicas was reconciling to the right MachineDeployments replicas in this scenario i.e 0.
This PR ensures the MachineDeployment replicas remains 0 all the time in the scenario above.