Use a cache-backed client to reduce number of api requests
We currently use an api reader which results in an actual api request
whenwever we need to get an object. This change makes us use the
default cache-backed client again. There are now a couple of conflict
errors in the log, but as that results in a retry it is not a
correctness issue.

Ref https://issues.redhat.com/browse/HOSTEDCP-154