Expose kubeadmin password in HostedCluster status
Before this PR, end users of Hypershift had no way to find the kubeadmin
password for their HostedCluster apart from looking inside the control
plane namespace.

This PR adds code to copy the kubeadmin password secret from the control
plane namespace to the HostedCluster's namespace and adds a field in the
status of the HostedCluster that references the password secret.