Replace contributors symlink with actual file
The b5aa26e0c8a0f69431fd727a21a749ebeaa3193d commit introduced a symlink
into the docs content, which breaks the local preview `make serve-containerized`
target (which throws file not found errors trying to deal with the link).

This commit replaces the symlink with an actual file to fix the regression.

We should consider having CONTRIBUTING.md just link to the docs page instead
of symlinking it.