docs: Revamp the docs and link to them from the README
This commit does some extensive overhauling of the docs site to improve
the navigation structure, add a call to action on the home page, clean
up a few content items, and add an initial private cluster guide.

This commit also strips most content out of the README and provides a
link to the latest published docs which is where all our content should
be maintained from now on.