Fix konnektivity-socks5-proxy
In 868bb057c4119b76ec3b4100ea88d55664c92715 I made the proxy close the
bodies it receives. This broke the proxy entirely, as closing the body
effectively discards any remaining data in there, but the data is the
response for the actual client which gets read by the proxy and then
forwarded.