Prevent flapping of nodepool config
Currently, we flap back and forth between different nodepool configs
that differ only in the user-data secret name. This happens because that
secret has a suffix that is the hash of it's configs. Those configs in
turn are assembled from a list of configmaps, but that list wasn't
sorted. Sort the list, be happy.