Default root-volume-size to 120 instead of 16 gig
While the 16 gig might be enough for a node to get into ready status, it
will very quickly get filled. Change this to use the same default as
self-hosted openshift. This will also improve the conformance test runs,
as we regularly have flakes there because all nodes are tainted because
of disk-pressure.