Infra destroy: Do not require cluster name for route 53 cleanup.
We currently require the name of the cluster, otherwise we don't clean
up the private route 53 zones. This is because they do not directly have
tags associated and can not be filtered by tags, so we filter by name
instead. This change makes us filter by VPC instead, which works withour
requiring the name.

Ref https://issues.redhat.com/browse/HOSTEDCP-287