PKI: Properly reconcile certificates and keys
Currently, we only check if a secret has the right entries for a
certificate and key but we never check on the content of those entries.
This means that for example if a certificate expires or its config
changes in any other way, we will not re-create it.

This change:
* Adds some code to verify a certificate and key against a config
* Uses this when reconciling certificates
* Cleans up the support/certs and
  hontrol-plane-operator/controllers/hostedcontrolplane/pki packages to
  not export methods that are not used in other packages

Ref https://issues.redhat.com/browse/HOSTEDCP-99