Convert hosted-cluster-config-operator into subcommand of CPO
This commit removes the discrete hosted-cluster-config-operator (HCCO) binary
and moves it to become a subcommand of control-plane-operator (CPO).

Architectural discussions have led to the conclusion that the logic of HCCO can
be incorporated into the CPO itself, since the programs are developed and
versioned in lockstep, coexist in the control plane namespace, and have very
similar scopes of access and responsibility. Eventually the HCCO could be part
of the CPO process itself, but this commit takes an incremental step towards
that end by relocating the code and removing the HCCO image.

There are two immediate benefits:

1. The control plane image that we incorporate into the OCP payload will become
   smaller (by shedding the 75MB HCCO binary).
2. The HCCO code will be closer to its final destination (a subpackage of CPO).

A future commit can merge HCCO logic into the CPO process and then remove all the
associated workload management code for HCCO (SA, role, deployment, etc.).