Fix update loops in hosted cluster config operator
Before this commit, the CPO was not passing the enable-ci-debug-output
flag to the hosted cluster config operator, so we were not detecting
update loops in CI. There were a few items causing update loops.

This PR enables loop dection for the hosted cluster config operator and
fixes resources that were causing update loops.