Reconcile in-cluster config for openshift apiserver and controller manager
Enables reconciliation of configuration that lives inside the cluster
with configuration that lives in the control plane for both the
openshift apiserver and the openshift controller manager.

This is for specific pieces of configuration such as Project, Build, and
Image config. Other than that, configuration should be specified on the
HostedCluster itself.