Move Ingress Operator into management cluster to speed up creation
This change moves the Ingress Operator into the management cluster in
order to speed up cluster creation. I did two testruns with this, in the
first it took 698 seconds, in the second 642 seconds to create a
cluster. This is about two minutes faster than the fastest run we
currently have from CI (14.2 minutes/852 seconds).