Add AWS Static credential suprt to some AWS functions when vendoring. infra.aws.create infra.aws.create_iam infra.aws.destroy infra.aws.destroy_iam
Signed-off-by: Joshua Packer <jpacker@redhat.com>

This allows the hypershift CLI infra AWS to work with key and secretKey credential values, when integrated into controllers or other environments where the credentials file may not be available.

This is to address issue: #859

This adds AWSKey and AWSSecretKey to the options we use for infra.aws.create and infra.aws.create_iam. The same was added to the destroy. This allows someone consuming the Go functions to include the credentials, but leaves the secret definition, access to the impelmenter