:warning: Move InfraID under AWS platform spec
This change moves the infra ID under the AWS platform spec, as it is an
AWS-specific concept. The InfraID was also used to name the CAPI
cluster, this was changed to instead name it like the hosted
controlplane as there will only be ever one capi cluster per HCP
namespace.

Ref https://issues.redhat.com/browse/HOSTEDCP-280