docs: Minor update to in-cluster development registry login steps
Before this commit, the in-cluster development docs used an `oc registry login`
command which could sometimes log the user in to the cluster's internal instead
of external registry if the prior steps to enable the external registry hadn't
yet completed.

This commit updates the login command to make sure the registry login always
targets the external registry route.