Crate cluster: Fix --wait panicing with a NPD
We changed the create command to Create the HostedCluster rather than
ServerSide applying it to make sure it errors if a HostedCluster of that
name already exists.

Unfortunately, the deserializer clears the TypeMeta field when doing
anything other than the serverside apply, so the code that looked at the
object kind after the Create to find the HostedCluster will never find
anything. Fix it by storing the HostedCluster before doing the Create
call.