Add a ClusterVersionSucceeding condition to HostedCluster
This commit adds a `ClusterVersionSucceeding` condition to HostedCluster which
tracks the `Failing` condition of the hosted cluster's `ClusterVersion`
resource. This condition provides additional context about the health of the
hosted cluster in the steady state and during release version rollouts. During a
rollout, information about individual operator rollout status will be surfaced
through this new condition.