Prevent accidental cluster deletion
Before this commit, the hosted cluster controller treated a missing
hostedcluster the same as a deleted one. This is incorrect, because any cache or
other transient weirdness from the apiserver for a hostedcluster which actually
still exists will result in that cluster being accidentally destroyed. Absence
of the hostedcluster being reconciled isn't the same as deletion intent.

This commit fixes the problem by short-circuiting reconciliation if the
requested hostedcluster doesn't exist. If on a subsequent event the resource can
be fetched, we'll reconcile it as usual. If we continue to get queued events for
a resource the apiserver says isn't found, that could be some other cache or
connectivity issue we want to surface and resolve. What we should not do is
infer a deletion event from it.