e2e: Reduce abstraction and fix teardown edge cases
Before this commit, the e2e code used a level of abstraction that significantly
obfuscated the fixture logic around create/teardown. This commit removes the
abstraction and brings the platform specific logic closer together to aid in
reader comprehension. As we get more experience with dealing with multiple
platforms, the emergent patterns should inform what abstractions make sense.
Right now we're only dealing with two, and they don't seem to differ in enough
ways to justify so much indirection.

As part of the refactor, this commit also fixes some teardown edge cases which
were only apparent after bringing the code closer together:

1. Cluster dumps during initial cluster teardown which failed but which
succeeded on retry used to incorrectly fail the test. Now if the dump eventually
suceeds, the test can succeed.

2. The cluster dump that occurs after test namespace teardown failure was
happening when the test failed for any reason. This is fixed so the dump only
happens if the namespace teardown fails.