Hostedcluster controller: Don't store log on Reconciler
The log is specific to each Reconcile() invocation, so we can not store
it persistently, as multiple concurrent Reconciles will stomp each
others logger.

This happens to not be a problem right now, because we run with only one
worker, but that is something we likely want to change to improve
throughput.