ensure hostedconfigoperator kube-controller-manager CA syncer uses combined ca to properly support ca rotation. Also ensure combined-ca set in pods service account's ca.crt field
Right now a pods service account is only getting created with one CA at a time: this will not work with CA rotation. The pod's serviceaccount CA field should be populated with the combined CA so that CA rotation can be properly supported (allows adding multiple CAs to the trust store at a given period of time)

Key changes:
In a pod with a service account
/var/run/secrets/kubernetes.io/ca.crt use to only have one root ca cert. Now it will include the combined ca bundle to ensure all active CAs are accounted for

The service-ca.crt in a pods service account use to only have the openshift-signer cert. It will now have the combinedCA bundle to ensure all CAs are accounted for