fix machineTemplate deletion logic to prevent panics in the hypershift operator (fixes IBMCloud deploys)
some providers like None and currently IBM Cloud UPI (Satellite) do not have explicit machine template types.

This PR changes that so the default is if something is not explicitly defined an empty list is returned which will allow a graceful skip of the machineTemplate deletion logic instead of a stopping cleanup of the nodepools.

Current code stops here:
https://github.com/openshift/hypershift/blob/767e7abff226c8559935a13b1287d377072e1633/hypershift-operator/controllers/nodepool/nodepool_controller.go#L150