add release image annotation support for extra metadata that can tie downstream deployments to a given release image
Currently there is no way to tie the deployments in the master-control plane to a given release. Obviously in the normal happy path: we assume over time things will reconcile and match the normal desired state of the cluster. There are also a few statuses that look at the operator versions in the cluster.

However: in a managed cloud offering: the internals of a user cluster can be in various states based on how they are operating the cluster including unhealthy states that might block the rollout of some downstream components (like monitoring operator) but not the managed control-plane side.

This pr adds an annotation that holds the release image a deployment was reconciled with. We can then look at a given deployment (like the kube-apiserver) and easily be able to reference it to the release image that deployed it. This helps with auditing and for additional tracking on rollouts of components