Set explicitOauthConfig as true when .OAuth is not nil
This PR https://github.com/openshift/hypershift/pull/787 intoduced a behaviour to prevent us reconciling the secret with user/pass for the guest cluster when the oauth config is set. However the boolean value passed through the reconciling function is wrongly inverted. This PR fixes it by seeting explicitOauthConfig as true when .OAuth != nil