Make JSON logging output in RFC3339 format
Before this commit, a Zap logging misconfiguration caused the RFC3339 timestamp
format option to be ignored because of the option being applied to the default
encoder instead of the user-supplied encoder being passed to `zap.New()`. As a
result, JSON logs were emitted with the EpochTimeEncoder.

This commit fixes all the logging configurations so that JSON logs use RFC3339
timestamps.