CLI: Simplify provider-specific kube resources
Currently, the provider has to implement an `AsObjects` interface that
returns a slice of objects. The only implementation of this interface is
for AWS which is assiged to a variable of the interface type, which
forces a type assertion two lines later.

All of this is needlessly complicated, we can simply pass a slice of
client objects from the provider-specific part into the genreric part
instead.