ensure all components respect the restart annotation.
This pr ensures all components respect the restart annotation. This is critical to support ca rotation because ignition server needs to reload the server certs after rotation which currently is not occuring. Additionally: it is opertionally expected from an ibm perspective that all components get restarted on a master refresh operation. This is an agreement we provide to customers that needs to be honored and is good operationally for all providers. There will be other situations in the future where an on demand restart is beneficial.