fix the node pull name in the kubevirt 2e2 test
The expected node pull name in the kubevirt test was changed to include AWS zone, where there is no AWS zone in kubevirt cluster, an so the test is fails because it can't find the node pull with the wrong name.

This PR reverts the naming change in the kubevirt test.

Signed-off-by: Nahshon Unna-Tsameret <nunnatsa@redhat.com>