Fix `install render` bugs
This commit fixes three issues with `install render`:

1. Before this commit, rendered install YAML output was broken because of a bug
causing the `---` delimiter to be omitted between the first two resources in
the list. This commit fixes the output to be valid YAML.

2. Before this commit, OIDC flag validation was incorrect and out of alignment
with the validation used for `install`, forcing users to specify the
`--oidc-storage-provider-s3-secret` when they are using the
`--oidc-storage-provider-s3-bucket-name` flag but no secret name. This commit
fixes the render code to reuse the existing correct validation code.

3. Before this commit, a regression caused `install render` to specify 0 replicas
for the hypershift-operator deployment regardless of the `--development` flag being
set or not. This commit fixes the issue by adding a consistent defaulting mechanism
shared by both `install` and `install render` so that the deployment once again
defaults to 1 replica unless the `--development` flag is specified.