Properly deal with hostedclusters that have a dot in the name
If a hostedcluster has a dot in its name, we currently fail to reconcile
it because we use the name as part of the namespace name and namespace
names do not allow dots.

Clusters with dots in their name end up being created by the periodic
jobs created by the release controller, because the release controller
uses the release as part of the prowjob name and we use the prowjob name
as stable identifier over steps to identify the cluster. Sample:
https://prow.ci.openshift.org/view/gs/origin-ci-test/logs/periodic-ci-openshift-hypershift-main-periodics-e2e-aws-pooled-periodic-conformance/1491754642944561152#1:build-log.txt%3A100