CLI: encrypt etcd by default
Adds flag `--aws-kms-key` to both `create iam aws` and
`create cluster aws` commands to allow specifying an existing KMS key to
use in encrypting etcd secrets. If this flag is specified, an IAM role
is created that allows the kms encryption plugin to use the key for
encryption/decryption.

If no KMS key is specified, a 32-byte AESCBC key is generated and
configured as the default encryption method.