Set the default AWS KMS provider image
The AWS KMS provider is now built by CI under the hypershift
organization. This sets up the image built by CI as the default image
used by the Kube API server when creating the KMS provider container.
The image can still be overriden by an annotation on the HostedCluster
resource.