Re-Enable guest cluster dump in e2e
We really want this to be able to debug e2e runs that fail due to
clusteroperators not becoming ready. I tried this locally, and it
consistently takes ~90 seconds to do a dump (n=3) which seems
reasonable. To be sure, I also added a log that will show how long it
took in CI.

This also exposes the guest dumping as a flag in the `hypershift dump`
command.