AWS: Use availability prober in capi provider
The AWS capi provider currently always errors for some time during
startup when a new cluster is created, because there is no token,
because the KAS is not up yet so the token minter sidecar can't mint
one. This change makes us use the availability prober init container to
avoid going into backoff due to this.

Using the availability prober required two new things to be available to
the provider:
* The HCP in case it uses a non-default apiserver port
* The Availability prober image

For the first, the existing provider interface was extended, as we only
have construct the hcp right before calling CAPIProviderDeploymentSpec.

For the static availability prober image we avoid extending the
interface for every provider for something that is only needed by AWS by
passing it into the provider at construction time.