Cluster create CMD: change code structure to have platform separation
Change cluster create CMD to more "Object Oriented" structure, in order to have
more separation between platform parametes and behavior.
Each platform parameters should be handled and managed in the platform cmd object
and not in the core cmd object.
This change also affect e2e tests, changed the e2e tests to use the new objects structure.
This change is done as preparation for the following changes:
1. Change API server default publishing strategy for NONE platforms (issue #1009)
2. Change Cluster create CMD to use NodePool create CMD code (issue #1010)